refactor: abstract updatePendingInvoice locked code

+++ core/api/src/app/wallets/update-single-pending-invoice.ts
@@ -36,6 +36,180 @@ import { toDisplayBaseAmount } from "@domain/payments"
 
 import { declineHeldInvoice } from "./decline-single-pending-invoice"
 
+const lockedUpdatePendingInvoiceSteps = async ({
+  paymentHash,
+  recipientWalletId,
+  receivedBtc,
+  description,
+  isSettledInLnd,
+  logger,
+}: {
+  paymentHash: PaymentHash
+  recipientWalletId: WalletId
+  receivedBtc: BtcPaymentAmount
+  description: string
+  isSettledInLnd: boolean
+  logger: Logger
+}) => {
+  const walletInvoices = WalletInvoicesRepository()
+
+  const walletInvoiceInsideLock = await walletInvoices.findByPaymentHash(paymentHash)
+  if (walletInvoiceInsideLock instanceof CouldNotFindError) {
+    logger.error({ paymentHash }, "WalletInvoice doesn't exist")
+    return false
+  }
+  if (walletInvoiceInsideLock instanceof Error) return walletInvoiceInsideLock
+  if (walletInvoiceInsideLock.paid) {
+    logger.info("invoice has already been processed")
+    return true
+  }
+
+  // Prepare metadata and record transaction
+  const recipientInvoiceWallet = await WalletsRepository().findById(recipientWalletId)
+  if (recipientInvoiceWallet instanceof Error) return recipientInvoiceWallet
+  const { accountId: recipientAccountId } = recipientInvoiceWallet
+
+  const accountWallets =
+    await WalletsRepository().findAccountWalletsByAccountId(recipientAccountId)
+  if (accountWallets instanceof Error) return accountWallets
+
+  const receivedWalletInvoice = await WalletInvoiceReceiver({
+    walletInvoice: walletInvoiceInsideLock,
+    receivedBtc,
+    recipientWalletDescriptors: accountWallets,
+  }).withConversion({
+    mid: { usdFromBtc: usdFromBtcMidPriceFn },
+    hedgeBuyUsd: { usdFromBtc: DealerPriceService().getCentsFromSatsForImmediateBuy },
+  })
+  if (receivedWalletInvoice instanceof Error) return receivedWalletInvoice
+
+  const {
+    recipientWalletDescriptor,
+    btcToCreditReceiver,
+    btcBankFee,
+    usdToCreditReceiver,
+    usdBankFee,
+  } = receivedWalletInvoice
+
+  addAttributesToCurrentSpan({
+    "invoices.finalRecipient": JSON.stringify(recipientWalletDescriptor),
+  })
+
+  if (!isSettledInLnd) {
+    const lndService = LndService()
+    if (lndService instanceof Error) return lndService
+    const invoiceSettled = await lndService.settleInvoice({
+      pubkey: walletInvoiceInsideLock.pubkey,
+      secret: walletInvoiceInsideLock.secret,
+    })
+    if (invoiceSettled instanceof Error) return invoiceSettled
+  }
+
+  const invoicePaid = await walletInvoices.markAsPaid(paymentHash)
+  if (invoicePaid instanceof Error) return invoicePaid
+
+  const recipientAccount = await AccountsRepository().findById(recipientAccountId)
+  if (recipientAccount instanceof Error) return recipientAccount
+  const { displayCurrency: recipientDisplayCurrency } = recipientAccount
+  const displayPriceRatio = await getCurrentPriceAsDisplayPriceRatio({
+    currency: recipientDisplayCurrency,
+  })
+  if (displayPriceRatio instanceof Error) return displayPriceRatio
+
+  const { displayAmount: displayPaymentAmount, displayFee } = DisplayAmountsConverter(
+    displayPriceRatio,
+  ).convert({
+    btcPaymentAmount: btcToCreditReceiver,
+    btcProtocolAndBankFee: btcBankFee,
+    usdPaymentAmount: usdToCreditReceiver,
+    usdProtocolAndBankFee: usdBankFee,
+  })
+
+  // TODO: this should be a in a mongodb transaction session with the ledger transaction below
+  // markAsPaid could be done after the transaction, but we should in that case not only look
+  // for walletInvoicesRepo, but also in the ledger to make sure in case the process crash in this
+  // loop that an eventual consistency doesn't lead to a double credit
+
+  const {
+    metadata,
+    creditAccountAdditionalMetadata,
+    internalAccountsAdditionalMetadata,
+  } = LedgerFacade.LnReceiveLedgerMetadata({
+    paymentHash,
+    pubkey: walletInvoiceInsideLock.pubkey,
+    paymentAmounts: {
+      btcPaymentAmount: btcToCreditReceiver,
+      usdPaymentAmount: usdToCreditReceiver,
+      btcProtocolAndBankFee: btcBankFee,
+      usdProtocolAndBankFee: usdBankFee,
+    },
+
+    feeDisplayCurrency: toDisplayBaseAmount(displayFee),
+    amountDisplayCurrency: toDisplayBaseAmount(displayPaymentAmount),
+    displayCurrency: recipientDisplayCurrency,
+  })
+
+  //TODO: add displayCurrency: displayPaymentAmount.currency,
+  const result = await LedgerFacade.recordReceiveOffChain({
+    description,
+    recipientWalletDescriptor,
+    amountToCreditReceiver: {
+      usd: usdToCreditReceiver,
+      btc: btcToCreditReceiver,
+    },
+    bankFee: {
+      usd: usdBankFee,
+      btc: btcBankFee,
+    },
+    metadata,
+    txMetadata: {
+      hash: metadata.hash,
+    },
+    additionalCreditMetadata: creditAccountAdditionalMetadata,
+    additionalInternalMetadata: internalAccountsAdditionalMetadata,
+  })
+  if (result instanceof Error) return result
+
+  // Prepare and send notification
+  const recipientUser = await UsersRepository().findById(recipientAccount.kratosUserId)
+  if (recipientUser instanceof Error) return recipientUser
+
+  const notificationResult = await NotificationsService().lightningTxReceived({
+    recipientAccountId,
+    recipientWalletId: recipientWalletDescriptor.id,
+    paymentAmount: receivedWalletInvoice.receivedAmount(),
+    displayPaymentAmount,
+    paymentHash,
+    recipientDeviceTokens: recipientUser.deviceTokens,
+    recipientNotificationSettings: recipientAccount.notificationSettings,
+    recipientLanguage: recipientUser.language,
+  })
+
+  if (notificationResult instanceof DeviceTokensNotRegisteredNotificationsServiceError) {
+    await removeDeviceTokens({
+      userId: recipientUser.id,
+      deviceTokens: notificationResult.tokens,
+    })
+  }
+
+  if (
+    recipientAccount.level === AccountLevel.One ||
+    recipientAccount.level === AccountLevel.Two
+  ) {
+    const callbackService = CallbackService(getCallbackServiceConfig())
+    callbackService.sendMessage({
+      accountUuid: recipientAccount.uuid,
+      eventType: CallbackEventType.ReceiveLightning,
+      payload: {
+        // FIXME: [0] might not be correct
+        txid: result.transactionIds[0],
+      },
+    })
+  }
+
+  return true
+}
+
 const updatePendingInvoiceBeforeFinally = async ({
   walletInvoice,
   logger,
@@ -48,15 +222,11 @@ const updatePendingInvoiceBeforeFinally = async ({
     pubkey: walletInvoice.pubkey,
   })
 
-  const lndService = LndService()
-  if (lndService instanceof Error) return lndService
-
   const walletInvoicesRepo = WalletInvoicesRepository()
 
   const {
     pubkey,
     paymentHash,
-    secret,
     recipientWalletDescriptor: recipientInvoiceWalletDescriptor,
   } = walletInvoice
 
@@ -73,6 +243,8 @@ const updatePendingInvoiceBeforeFinally = async ({
     onUs: false,
   })
 
+  const lndService = LndService()
+  if (lndService instanceof Error) return lndService
   const lnInvoiceLookup = await lndService.lookupInvoice({ pubkey, paymentHash })
   if (lnInvoiceLookup instanceof InvoiceNotFoundError) {
     const isDeleted = await walletInvoicesRepo.deleteByPaymentHash(paymentHash)
@@ -84,15 +256,15 @@ const updatePendingInvoiceBeforeFinally = async ({
   }
   if (lnInvoiceLookup instanceof Error) return lnInvoiceLookup
 
-  const {
-    lnInvoice: { description },
-    roundedDownReceived: uncheckedRoundedDownReceived,
-  } = lnInvoiceLookup
-
   if (walletInvoice.paid) {
     pendingInvoiceLogger.info("invoice has already been processed")
     return true
   }
+
+  const {
+    lnInvoice: { description },
+    roundedDownReceived: uncheckedRoundedDownReceived,
+  } = lnInvoiceLookup
 
   if (!lnInvoiceLookup.isHeld && !lnInvoiceLookup.isSettled) {
     pendingInvoiceLogger.info("invoice has not been paid yet")
@@ -116,163 +288,16 @@ const updatePendingInvoiceBeforeFinally = async ({
   if (receivedBtc instanceof Error) return receivedBtc
 
   const lockService = LockService()
-  return lockService.lockPaymentHash(paymentHash, async () => {
-    // we're getting the invoice another time, now behind the lock, to avoid potential race condition
-    const invoiceToUpdate = await walletInvoicesRepo.findByPaymentHash(paymentHash)
-    if (invoiceToUpdate instanceof CouldNotFindError) {
-      pendingInvoiceLogger.error({ paymentHash }, "WalletInvoice doesn't exist")
-      return false
-    }
-    if (invoiceToUpdate instanceof Error) return invoiceToUpdate
-    if (invoiceToUpdate.paid) {
-      pendingInvoiceLogger.info("invoice has already been processed")
-      return true
-    }
-
-    // Prepare metadata and record transaction
-    const recipientInvoiceWallet = await WalletsRepository().findById(
-      recipientInvoiceWalletDescriptor.id,
-    )
-    if (recipientInvoiceWallet instanceof Error) return recipientInvoiceWallet
-    const { accountId: recipientAccountId } = recipientInvoiceWallet
-
-    const accountWallets =
-      await WalletsRepository().findAccountWalletsByAccountId(recipientAccountId)
-    if (accountWallets instanceof Error) return accountWallets
-
-    const receivedWalletInvoice = await WalletInvoiceReceiver({
-      walletInvoice: invoiceToUpdate,
+  return lockService.lockPaymentHash(paymentHash, async () =>
+    lockedUpdatePendingInvoiceSteps({
+      recipientWalletId: recipientInvoiceWalletDescriptor.id,
+      paymentHash,
       receivedBtc,
-      recipientWalletDescriptors: accountWallets,
-    }).withConversion({
-      mid: { usdFromBtc: usdFromBtcMidPriceFn },
-      hedgeBuyUsd: { usdFromBtc: DealerPriceService().getCentsFromSatsForImmediateBuy },
-    })
-    if (receivedWalletInvoice instanceof Error) return receivedWalletInvoice
-
-    const {
-      recipientWalletDescriptor,
-      btcToCreditReceiver,
-      btcBankFee,
-      usdToCreditReceiver,
-      usdBankFee,
-    } = receivedWalletInvoice
-
-    addAttributesToCurrentSpan({
-      "invoices.finalRecipient": JSON.stringify(recipientWalletDescriptor),
-    })
-
-    if (!lnInvoiceLookup.isSettled) {
-      const invoiceSettled = await lndService.settleInvoice({ pubkey, secret })
-      if (invoiceSettled instanceof Error) return invoiceSettled
-    }
-
-    const invoicePaid = await walletInvoicesRepo.markAsPaid(paymentHash)
-    if (invoicePaid instanceof Error) return invoicePaid
-
-    const recipientAccount = await AccountsRepository().findById(recipientAccountId)
-    if (recipientAccount instanceof Error) return recipientAccount
-    const { displayCurrency: recipientDisplayCurrency } = recipientAccount
-    const displayPriceRatio = await getCurrentPriceAsDisplayPriceRatio({
-      currency: recipientDisplayCurrency,
-    })
-    if (displayPriceRatio instanceof Error) return displayPriceRatio
-
-    const { displayAmount: displayPaymentAmount, displayFee } = DisplayAmountsConverter(
-      displayPriceRatio,
-    ).convert({
-      btcPaymentAmount: btcToCreditReceiver,
-      btcProtocolAndBankFee: btcBankFee,
-      usdPaymentAmount: usdToCreditReceiver,
-      usdProtocolAndBankFee: usdBankFee,
-    })
-
-    // TODO: this should be a in a mongodb transaction session with the ledger transaction below
-    // markAsPaid could be done after the transaction, but we should in that case not only look
-    // for walletInvoicesRepo, but also in the ledger to make sure in case the process crash in this
-    // loop that an eventual consistency doesn't lead to a double credit
-
-    const {
-      metadata,
-      creditAccountAdditionalMetadata,
-      internalAccountsAdditionalMetadata,
-    } = LedgerFacade.LnReceiveLedgerMetadata({
-      paymentHash,
-      pubkey: walletInvoice.pubkey,
-      paymentAmounts: {
-        btcPaymentAmount: btcToCreditReceiver,
-        usdPaymentAmount: usdToCreditReceiver,
-        btcProtocolAndBankFee: btcBankFee,
-        usdProtocolAndBankFee: usdBankFee,
-      },
-
-      feeDisplayCurrency: toDisplayBaseAmount(displayFee),
-      amountDisplayCurrency: toDisplayBaseAmount(displayPaymentAmount),
-      displayCurrency: recipientDisplayCurrency,
-    })
-
-    //TODO: add displayCurrency: displayPaymentAmount.currency,
-    const result = await LedgerFacade.recordReceiveOffChain({
       description,
-      recipientWalletDescriptor,
-      amountToCreditReceiver: {
-        usd: usdToCreditReceiver,
-        btc: btcToCreditReceiver,
-      },
-      bankFee: {
-        usd: usdBankFee,
-        btc: btcBankFee,
-      },
-      metadata,
-      txMetadata: {
-        hash: metadata.hash,
-      },
-      additionalCreditMetadata: creditAccountAdditionalMetadata,
-      additionalInternalMetadata: internalAccountsAdditionalMetadata,
-    })
-    if (result instanceof Error) return result
-
-    // Prepare and send notification
-    const recipientUser = await UsersRepository().findById(recipientAccount.kratosUserId)
-    if (recipientUser instanceof Error) return recipientUser
-
-    const notificationResult = await NotificationsService().lightningTxReceived({
-      recipientAccountId,
-      recipientWalletId: recipientWalletDescriptor.id,
-      paymentAmount: receivedWalletInvoice.receivedAmount(),
-      displayPaymentAmount,
-      paymentHash,
-      recipientDeviceTokens: recipientUser.deviceTokens,
-      recipientNotificationSettings: recipientAccount.notificationSettings,
-      recipientLanguage: recipientUser.language,
-    })
-
-    if (
-      notificationResult instanceof DeviceTokensNotRegisteredNotificationsServiceError
-    ) {
-      await removeDeviceTokens({
-        userId: recipientUser.id,
-        deviceTokens: notificationResult.tokens,
-      })
-    }
-
-    if (
-      recipientAccount.level === AccountLevel.One ||
-      recipientAccount.level === AccountLevel.Two
-    ) {
-      const callbackService = CallbackService(getCallbackServiceConfig())
-      callbackService.sendMessage({
-        accountUuid: recipientAccount.uuid,
-        eventType: CallbackEventType.ReceiveLightning,
-        payload: {
-          // FIXME: [0] might not be correct
-          txid: result.transactionIds[0],
-        },
-      })
-    }
-
-    return true
-  })
+      isSettledInLnd: lnInvoiceLookup.isSettled,
+      logger,
+    }),
+  )
 }
 
 export const updatePendingInvoice = wrapAsyncToRunInSpan({
@@ -291,8 +316,9 @@ export const updatePendingInvoice = wrapAsyncToRunInSpan({
     })
     if (result) {
       if (!walletInvoice.paid) {
-        const walletInvoices = WalletInvoicesRepository()
-        const invoicePaid = await walletInvoices.markAsPaid(walletInvoice.paymentHash)
+        const invoicePaid = await WalletInvoicesRepository().markAsPaid(
+          walletInvoice.paymentHash,
+        )
         if (
           invoicePaid instanceof Error &&
           !(invoicePaid instanceof CouldNotFindWalletInvoiceError)
